NMS-8773: Migrate existing EnhancedLinkdTopology to make use of Layers
