+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -33,11 +33,6 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
     <reference id="cdpElementDao" interface="org.opennms.netmgt.dao.api.CdpElementDao" availability="mandatory"/>
     <reference id="ipNetToMediaDao" interface="org.opennms.netmgt.dao.api.IpNetToMediaDao" availability="mandatory"/>
 
-    <bean id="enlinkdTopologyProviderInfo" class="org.opennms.features.topology.api.topo.DefaultTopologyProviderInfo">
-        <property name="name" value="Enhanced Linkd Topology Provider"/>
-        <property name="description" value="This Topology Provider displays the topology information discovered by the Enhanced Linkd daemon. It uses the SNMP information of several protocols like OSPF, ISIS, LLDP and CDP to generate an overall topology."/>
-    </bean>
-
     <!-- Metrics -->
     <bean id="metricRegistry" class="com.codahale.metrics.MetricRegistry" />
 
@@ -61,9 +56,8 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
     </bean>
     <!-- Shared Components end -->
 
-    <bean id="topologyProvider" scope="singleton" class="org.opennms.features.topology.plugins.topo.linkd.internal.EnhancedLinkdTopologyProvider">
-        <argument ref="metricRegistry" />
-        <property name="configurationFile" value="${configurationFile}"/>
+    <bean id="topologyFactory" scope="singleton" class="org.opennms.features.topology.plugins.topo.linkd.internal.TopologyFactory">
+        <property name="metricRegistry" ref="metricRegistry" />
         <property name="transactionOperations" ref="transactionOperations"/>
         <property name="nodeDao" ref="nodeDao" />
         <property name="snmpInterfaceDao" ref="snmpInterfaceDao" />
@@ -81,9 +75,9 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
         <property name="bridgeMacLinkDao" ref="bridgeMacLinkDao" />
         <property name="bridgeBridgeLinkDao" ref="bridgeBridgeLinkDao" />
         <property name="bridgeTopologyDao" ref="bridgeTopologyDao" />
-        <property name="topologyProviderInfo" ref="enlinkdTopologyProviderInfo"/>
         <property name="ipNetToMediaDao" ref="ipNetToMediaDao"/>
     </bean>
+
 
     <!--
       Wrap the Linkd provider in a provider that calculates SZL based on hop distance from a set of
@@ -91,36 +85,19 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
     -->
     <reference id="enlinkd" interface="org.opennms.netmgt.model.ServiceDaemon" filter="(name=enlinkd)" availability="optional"/>
 
-    <bean id="topologyHopProvider" class="org.opennms.features.topology.api.support.VertexHopGraphProvider">
-        <argument ref="topologyProvider"/>
-    </bean>
-
-    <bean id="topologyACLProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.NodeACLVertexProvider">
-        <argument ref="topologyHopProvider" />
-        <argument ref="nodeDao" />
-    </bean>
-
-    <!-- Linkd Topology Provider Services -->
-    <service ref="topologyACLProvider">
-    	<interfaces>
-    		<value>org.opennms.features.topology.api.topo.VertexProvider</value>
-    		<value>org.opennms.features.topology.api.topo.EdgeProvider</value>
-    		<value>org.opennms.features.topology.api.topo.GraphProvider</value>
-    	</interfaces>
-    </service>
-
     <!-- Meta topology provider -->
-    <bean id="enhancedLinkdMetaTopologyProvider" class="org.opennms.features.topology.api.topo.SimpleMetaTopologyProvider">
-        <argument ref="topologyACLProvider" />
+    <bean id="enhancedLinkdMetaTopologyProvider"
+          class="org.opennms.features.topology.api.topo.SimpleMetaTopologyProvider"
+          factory-ref="topologyFactory"
+          factory-method="createMetaTopologyProvider">
     </bean>
-
     <service interface="org.opennms.features.topology.api.topo.MetaTopologyProvider" ref="enhancedLinkdMetaTopologyProvider">
         <service-properties>
             <entry key="label" value="Enhanced Linkd" />
         </service-properties>
     </service>
 
-    <service interface="org.opennms.features.topology.api.topo.SearchProvider" ref="topologyProvider" />
+    <!--<service interface="org.opennms.features.topology.api.topo.SearchProvider" ref="topologyProvider" />-->
 
     <!-- EnhancedLinkd StatusProviders Service and CheckedOperation -->
     <bean id="linkdEdgeStatusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdEdgeStatusProvider">
