+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/test/resources/META-INF/opennms/applicationContext-enhanced-mock.xml
@@ -73,7 +73,7 @@
 
     <bean id="metricRegistry" class="com.codahale.metrics.MetricRegistry" />
 
-    <bean id="enLinkdTopologyProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.EnhancedLinkdTopologyProvider">
+    <bean id="enLinkdTopologyProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.providers.EnhancedLinkdTopologyProvider">
         <constructor-arg ref="metricRegistry"/>
         <property name="bridgeTopologyDao" ref="bridgeTopologyDao" />
         <property name="bridgeBridgeLinkDao" ref="bridgeBridgeLinkDao" />
